[DM-23942] Rewrite using aiohttp and Safir
Replace Flask with aiohttp and Safir, using some aiohttp addons to
provide functionality that Flask was previously handling.  This
requires replacing flash messages with manually passing a message in
the session, and doing some more manual work for CSRF handling.

Refactor the handlers to follow the Roundtable aiohttp bot pattern,
and rename the tests accordingly.

Sessions are stored in an encrypted cookie, using a key that's reset
on service restart.  This should be fine for now, although this may
be refactored to use a persistent key in the future.