[DM-23776] Change algorithm for GitHub groups
Instead of using the organization, a colon, and the group name,
generate UNIX-friendly group names.  The default will be the
organization name, a dash, and the team slug.  If longer than 32
characters, this will be truncated and made unique.

Add the team ID as an id attribute in the isMemberOf claim for the
generated groups.

Add documentation for the GitHub token generation and the
algorithm for generating the group name.