[DM-23776] Prefer session auth to Authorization
JupyterHub sends an Authorization header on all AJAX requests, but
we were prefering the Authorization header to a session cookie if
one were present.  Therefore, the /auth endpoint was trying to
parse the JupyterHub Authorization header, failing, and acting as
if the user had no authentication.

Prefer the session if one exists, and only then fall back on the
Authorization header.