[DM-23877] Add new abstractions for tokens and providers
Add an abstract base class for authentication providers and move
more of the GitHub-specific logic into GitHubProvider.

Add a new abstraction for tokens that providers use, and clean up
lots of code that was passing around dicts and strs as tokens to
use the new abstraction, which provides additional type checking.

Add missing files to the documentation build and clean up some
errors that uncovered.