[DM-23877] Refactor test suite
Break the huge tests.util class into a bunch of smaller modules.
Create a test configuration class to capture details about the test
environment, stash it in the test application, and retrieve it from
the test application when necessary in tests.

Properly mock aiohttp.ClientSession and get rid of the class
structure that was designed to allow mocking of the HTTP client
pieces of various classes.

Change the upstream issuer URL to be upstream.example.com to make
some of the examples more straightforward.