[DM-23877] Overhaul session handling
Now that OpenID Connect is implemented internally, there is no need
to remain compatible with oauth2_proxy sessions.  Completely rewrite
the session handling and the token store.

Sessions are now encrypted with Fernet using the same key as the
session cookie.  The returned session handle that the user can use
in place of a token has two parts, a key and a secret.  The key is
the identifier used for the session handle in the /auth/token routes.
The secret is used to verify ownership and to prevent someone who can
see the Redis keys from using them as handles.

Encoded handles with their secrets have a prefix, as do the session
keys in Redis.

The token store now stores index entries for tokens instead of the
full claim set, and expiring index entries is now a separate call
instead of a side effect of getting a list of tokens.