[DM-23877] Simplify token verification
Internally-issued tokens don't need to go through the full
verification process that retrieves a JWKS via an HTTP client.  We
already have the key and know what issuer to expect, so we can have
the issuer verify the token directly.  This in turn removes support
for tokens from upstream issuers from internal routes, which is what
we want anyway.

Full token verification is now supported only by the OIDC provider,
since it's the only component that needs to talk to an external
issuer.

As part of this change, store an RSAKeyPair object in the issuer
configuration instead of just the private key in PEM format so that
we have easier access to the public key.  Simplify the well-known
route handler, since it doesn't need to construct an RSAKeyPair now.

Move the ALGORITHM constant into a new constants module to avoid
circular dependencies.