[DM-23877] Additional verification of token claims
Require that all tokens have username and uid claims (however
configured) and reject ones that do not have those claims.  Lift
some key claims we care about into a VerifiedToken for ease of
access, and be stricter internally about using a jti.

Do some additional refactoring of token verification to reunify
all of it in the TokenVerifier class, and move token analysis into
that class as well.  This undoes a bit of the code moves from an
earlier diff after further reflection.