[DM-23877] Support session handles in Authorization
In various refactorings, support for session handles in the
Authorization header was accidentally dropped.  Reintroduce that
support.  Change the prefix for session handles to gsh- instead of
gsh: for compatibility with HTTP Basic Auth.

Refactor the handler helpers to move code into the modules that
use it and simplify some of the code.

Stop returning headers on forbidden errors.  I can't imagine
anything will bother using them.