[DM-23877] Protect against open redirects
Require that the redirect URL provided to the login handler is at
the same host as the login route.  This is necessary anyway because
authentication state is stored in a host-bound cookie.

Explicitly document this, and rename the Kubernetes documentation
to installation since that's the only supported installation method.

Fix a few bugs in the install documentation that I noticed while
editing it.