[DM-23877] Move test configuration to the test suite
Move the test configuration out of defaults.yaml to a separate
YAML file in the test suite and load it.  Use several environments
in that configuration file rather than passing parameters directly
into LazySettings.

Base a few more tests on the value of the configuration setting so
that the settings values aren't duplicated in several places.