[DM-23877] Restructure group_mappings internally
It's better for the code if group_mappings maps a group to a set
of scopes that group provides, so reconstruct it during config
parsing into that form.  Move group_mapping into the issuer config
since that's the only place it's now used.

Fix display of the scopes when logging successful or failed
authorization.