[DM-23877] Hide more test setup complexity
Use a fixture to encapsulate the process of creating a test server
and a test client.  This is a bit more complex than it otherwise
would be since we want to take a parameter, but this approach still
hides fixtures the test doesn't need to know about and simplifies
the code a bit.

Change a few more tests to use configuration values rather than
hard-coding things.