[DM-23943] Use pydantic instead of Dynaconf
Dynaconf is a little too weird and difficult to deal with.  Replace
it with pydantic and use a single configuration file instead of
layered configuration files.

Remove the logic to pass in custom configuration settings and instead
use a directory of test setting files that are processed as templates.

Turn some of the Config settings into tuples to make them immutable,
and change the group_mapping values to frozensets for the same reason.