[DM-24635] Overhaul authentication challenges
Allow the authentication type for the challenge in WWW-Authenticate
to be configured via the auth_type parameter to the /auth route.
Some applications will want to present a Basic challenge instead.

Bring the challenges into compliance with RFC 6750 when returning
a Bearer challenge.  This specifically means no longer using a
free-form error and instead using the registered error types,
returning 400 instead of 401 in cases where the header could not
be parsed, and including a challenge with 400 and 403 responses.

Separate authenticated route handling for the tokens web interface
from the /auth handling.  The former only needs to know about the
API exported by ingress-nginx with the /auth handler, and mixing
that into the full parsing and return codes for the /auth handler
was confusing and fragile.  This resulted in a lot of code
reshuffling to move code only used from one module into that module.

Wrap authentication challenges in some data types to make the code
a bit more robust and readable.