[DM-24685] Pass loggers into the ComponentFactory
In preparation for making more extensive use of structlog, fix
the typing on the logger objects so that the code is aware they
are structlog loggers and pass the logger explicitly into factories
rather than implicitly via the request.  The latter is because
binding additional context returns a modified logger, and we want
to support passing that modified logger into other components.

Fix the ordering of the Safir logging middleware and aiohttp_remotes
so that in a future world in which Safir adds remote IP information
directly, it will get the benefit of the aiohttp_remotes
transformations.