[DM-24930] Drop support for Python 3.7
Debian unstable no longer contains a working Python 3.7 and it was
previously required to update dependencies because of the
importlib-metadata versioned dependency.  Drop support for
Python 3.7 for simplicity.  Remove workarounds for Python 3.7 in
various places and change the Docker image to use Python 3.8.

Regenerate dependencies, which picks up various version updates.