[DM-24930] Refactor ClientFactory handling
Push more of the logic down into the RequestContext and make the
ClientFactory simpler.  This removes all the parameters to the
factory methods and shifts the burden of acquiring all the objects
to the construction of the factory, which makes the code a bit
more straightforward and easier to read.