[DM-24930] Clean up logging
Be a bit more consistent about what level things are logged as.
Log the configured trusted proxy IPs on startup.  Log an info-level
starting message, and ensure there is reasonable logging if the
log level is raised to info from debug.  Improve the logging of
token verification from an OpenID Connect provider.