[DM-24930] Allow tests to override settings
Allow overrides for settings to be passed in as extra parameters
to create_app, and plumb that through the test fixtures so that
tests can make use of it.  Use this to get rid of the github-claims
environment, which was just a copy of the github environment with
a couple of settings overrides.