[DM-24930] Better test token verification
Add some tests for a lot of error cases for token verification.
Fix unhandled exceptions for a missing key ID in the header of a
token and when the key ID doesn't appear in the JWKS.  Change the
exceptions returned by verify_oidc_token to be a little more
sensible.