Initial implementation of an OpenID Connect server
Add an OpenID Connect server that, rather than using an nginx
auth_request handler in front of a service, allows the service
to use the OpenID Connect protocol to request user authentication
and get authentication details.

Configuration for this is done via a new configuration option that
points to a file of JSON-encoded client IDs and secrets for valid
clients.