Refactor session storage handling
Abstract the Redis storage layer into a separate generic class
from the SessionStore in preparation for reusing the storage code
for other storage requirements.

This requires introducing an intermediate representation of a
session, SerializedSession, because creating a Session requires
additional objects and additional information not contained in the
storage.  SessionStore handles the conversion.