Rename TokenStore to UserTokenStore
Rename the class to be clearer about what it's storing, and move
the containing module under gafaelfawr.storage.  Derive
UserTokenEntry from Serializable to use some of the default
documentation.

Move the encoded representation to a hidden variable of
UserTokenEntry that's handled as an InitVar and is used by
to_json instead of being referenced as an attribute.  This makes
the code a little cleaner and hides information a little better.