Move exception serialization logic into exceptions
For the exceptions that can be returned by the redeem_code method
of an OIDCServer, put them into an exception hierarchy and add some
helper methods for logging and error reporting.  Make use of those
in the handler.

Rename the exceptions to end in Error instead, mostly because it's
shorter but partly because it matches the RFC terminology.