Improve error handling from /auth/openid/login
Reading the OAuth 2.0 specification in more detail, it says that
a request with an invalid client_id should not be redirected back
to the redirect_uri.  Replace that logic with a straight HTTP
error.

Refactor the construction of the return URL after login into a
helper function, which will be useful later when other errors
generate redirects.