Refactor OpenID authentication request parsing
Check the client_id first, with a new method in the OpenID Connect
server class, since if it isn't valid we are not supposed to report
errors via redirect.

Move the parsing of the rest of the request into a helper class and
use rich exceptions to make building a return URL easier.

Improve the return URL abstraction to take the new authentication
request class and automatically include the state parameter.