Unify helpers for unauthorized challenges
Add a separate helper function to return an unauthorized challenge
that optionally changes the HTTP status code if the request was
sent with AJAX.  Use that function in the /auth handler and in all
the places that unauthorized errors with WWW-Authenticate headers
were returned.

The previous AuthChallenge construction made it possible to define
an error without an error_description, which then required code to
handle that special case that was never exercised (because it's an
error).  Restructure AuthChallenge into a class hierarchy
representing the two types of challenges to eliminate this potential
error, and adjust tests accordingly.