Further error handling cleanups
Add two more exceptions that represent more OAuth standard error
codes, and generate them on the fly to use the standard machinery
for logging and generating error replies.  Add logging to some
error paths that didn't log the error properly and try to ensure
that all errors are logged with an error context key.  Remove the
now-unnecessary forbidden helper function and inline the challenge
generation for the /auth/forbidden handler since it's a special
case.