Clarify OpenID Connect authentication flow
In the diagram, do not include the URL to which the browser is
being redirected in the redirect label, since it makes the labels
too long and they are truncated.  Instead, just note a redirect,
since the next line shows the resulting URL.

Explicitly note that the redirect back from /login sets a session
cookie, and note this in the flow description.