Add FastAPI implementation
Add a parallel implementation of Gafaelfawr using FastAPI instead
of aiohttp.  Add a parallel set of route handler tests for that
implementation.

This new implementation is currently identical to the previous
implementation except that none of the handlers support being
protected by Gafaelfawr itself and retrieving a token from an HTTP
header.  Instead, they must be used either with Authorization
headers or with a cookie-based session.