Add Selenium testing for FastAPI
This required undoing the work to switch the /auth/token endpoints
over to using sessions instead of tokens, since there is currently
no way to simulate a session with Selenium.  Add support for the
X-Auth-Request-Token header for endponts protected by verify_token.

There are some ugly hacks here for Redis initialization and shutdown
that will be fixed in later test cleanup.