Convert to httpx for the HTTP client
This requires replacing aioresponses with the (much inferior)
pytest-httpx, which in turn requires a lot of restructuring of
the callbacks and registered responses.  Also get rid of a few
more remnants of aiohttp sitting around, although not yet all of
it pending some final work on the X-Forwarded middleware.

The aiohttp test harness was providing automagical support for
async tests.  Replace that with the explicit pytest-asyncio marks.