Make XForwardedMiddleware more generic
Avoid a direct dependency on configuration loading and instead take
the trusted hosts as a configuration parameter to the constructor.
Allow this to work with the test suite by loading middleware during
the startup hook instead of at module load time.  That in turn
requires using asgi-lifespan to send the correct lifecycle events,
since httpx's AsyncClient doesn't do this.

Remove the old aiohttp XForwardedFiltered middleware and port all of
its tests to the new Starlette middleware.