Add manager for token administrators
Plumb the database creation all the way through Gafaelfawr including
the settings, add a command-line utility to initialize the database,
add testing machinery to set up a database, and fix a few problems
with the database schema turned up by this process.

Add an AdminManager class that can add and retrieve token
administrators and a simple test suite, as well as the plumbing
through to storage classes.