Add user-info and token-info routes
Implement user-info and token-info routes using the new token
storage.  Refactor the token models to share more data and to use
the JSON encoding built into Pydantic instead of rolling our own.
Get rid of the remaining dataclasses and be consistent about using
Pydantic models for token data.  Include the token in the stored
Redis token data, which simplifies a lot of the argument passing
and parsing.