Install dependencies to build psycopg2
The psycopg2 documentation recommends not using the prebuilt
binaries for production builds, so add the necessary dependencies
to build it to the Docker image.

In the long run, these should move to a different layer so that we
can remove the compilers and other machinery in the production
container (to make it smaller).