Remove old user token UI
This is replaced by the new API and a yet-to-be-written JavaScript
UI frontend.  The type of tokens that this creates will be going
away, replaced with opaque tokens, so remove the entire UI rather
than port it.

Clean up CSRF handling so that it isn't tied to authentication and
adjust the test suite accordingly.

As of this commit, there are no more Selenium tests, but they'll be
reintroduced once the JavaScript UI is added.