Remove all internal JWT support
Switch to opaque tokens everywhere and remove support for internal
JWT tokens except for the small OpenID Connect server.  Delete all
the remaining session code in favor of the new token storage code.

Move the old Token and VerifiedToken models into an OIDC-specific
models file.

Rework the analyze route to make it meaningful with the new tokens.
The layout of the JSON it returns is not 100% backwards compatible.