Add support for internal and notebook tokens
Stop providing the downstream service with the token by default
from the /auth endpoint.  Instead, require that it request a token
with either the notebook or the delegate_to/delegate_scope query
parameters.  Add support for creating those types of tokens.

This doesn't yet reuse tokens, so isn't deployable as is since it
would flood the token database with tokens.