Clean up permission handling
Push more of the permission checks for token operations down into
the service and add tests for that functionality.  Return 404
instead of 403 when referring to tokens with the wrong user, and add
more machinery to detect this case.  Add logging to the token service
for interesting events (token creation, deletion, modification, and
permission denied errors).