Fix user token expiration and name handling
Allow for creating user tokens that don't expire and modifying an
existing user token to not expire.  (Note that this currently does
not affect child tokens, and they will still be reused even if they
are about to expire.)

Creating tokens with duplicate names was already handled in the
database via a constraint, but the error wasn't handled and resulted
in an internal server error.  Check for name duplication before
creating or renaming a token and return a properly handled exception
and a correct HTTP error.