Fix dev server with FastAPI
Add PostgreSQL to the docker-compose file.  Fix initialization of
PostgreSQL by adding SQLite-specific parameters only if the
database DSN is SQLite.  Make the port numbrer for the dev server
consistent.  Change the environment variable for setting the
Gafaelfawr configuration to GAFAELFAWR_SETTINGS_PATH and rename
CONFIG_PATH to SETTINGS_PATH internally to be more consistent.