Update Docker for FastAPI
Use the recommended FastAPI image and add a prestart script that
initializes the database (eventually to use Alembic).  Add a test
docker-compose configuration to start a local server with Docker
instead of outside of it.

Add some additional logging to the database setup so that there
are logs from the database initialization, and move configuration
of logging to the configuration file loading.