Create tokens in PostgreSQL first and then Redis
Redis is the primary store used for authentication and the point
at which the token becomes "active."  If the token exists in
PostgreSQL but not Redis, the worst consequence is some user
confusion, but a token in Redis but not PostgreSQL may not be
usable.  So create the token in PostgreSQL first, then Redis.