Add top-level error handling
Capture errors and display them in a div at the top-level page
instead of sending them to the console.  Replace errors from fetch
network problems with a more meaningful error message, and parse
the error messages from Gafaelfawr.

This is still not perfect in that it doesn't correctly handle
Gafaelfawr error messages that aren't in the JSON format that we
expect, but it's a good start.

Also declare the Content-Type when posting data to the backend.