Return 401 instead of redirects to API calls
The React frontend prefers a 401 error because that's easier to
handle and turn into appropriate action.  Redirects on the AJAX
calls are harder to deal with.  Therefore, add a new dependency
that requires a session cookie but returns 401 instead of a redirect
and use it for API calls.

Plumb through ajax_forbidden as a parameter to translate 401 into
403 and use it only for the /auth route, not for the API routes.