Add ESLint and fix errors
Add configuration for ESLint using eslint-config-wesbos and fix
all of the problems it found other than the ones involving typing.
This required moving LoginContext to its own file and wrapping
the callbacks in useCallback so that the new dependencies on them
didn't trigger continuous reloading.

Weirdly, this seemed to fix the focus problem with the token
creation modal.  I'm not sure why.