Refactor authentication dependency
Turn the various autentication and authorization dependencies into
a class with a variety of parameters to unify the code and adjust
all of the handlers accordingly.  Represent token admin rights with
a scope rather than just a database table, and add that scope to
users as they authenticate if they are an admin.

This plumbs through the parameter for allow_bootstrap but it doesn't
do anything yet.