Unify CSRF and authentication dependencies
The CSRF checking should only apply when authentication is done
via a browser session cookie, so implementing it as a separate
dependency was awkward since it didn't know which version of the
authentication dependency to depend on.  Instead, integrate CSRF
verification with authentication checking and trigger it
automatically based on the HTTP method so that no one has to
remember to add the verify_csrf dependency.

Only the login route should ever set the CSRF state in the cookie,
so move the set_csrf dependency code there.