Add documentation of bootstrap and service tokens
This documentation is somewhat incomplete since it doesn't include
Helm or secret changes, since the Helm chart has not yet been
updated for Gafaelfawr 2.0.0.  That will be done as part of the
2.0.0 release process.

Add a make docs target that filters out the errors from Sphinx for
the packages that don't have Sphinx-compatible documentation.