Update dependencies
Pin the version of idna and add a spurious dependency on aiohttp to
avoid version conflicts between requirements/dev.txt and
requirements/main.txt.

Update typing for changes to pytest, structlog, and jwt.  jwt is
very confused about the type of key information for RSA algorithms,
but passing in a PEM as a str seems to work.

Pin the version of selenium-wire prior to version 4, since version
4 breaks tests that are run async.