Clean up models and token service
Add some more default factories and validators to the token and
history models to reduce code a little.  Abstract normalize_scopes
and current_datetime into utility functions.  Change the database
representation of no scopes to be the empty string to be more
consistent with the history old_scopes entry.  Make a few more
parameters required, and require an empty list for scopes in more
places rather than allowing None.