Add a per-user token change history search form
Refactor the UI to add a common layout component that captures the
login context and will eventually handle some other things like
page styling and the error banner.  Add a UI to search the token
change history for a username.  Reuse the table component of it for
the information page for an individual token.