Change /auth/analyze output to be backward compatible
Undo a poor choice I made during the FastAPI rewrite to change the
output format of /auth/analyze.  Nublado JupyterHub is using this
for now to retrieve token metadata, so convert the new token data
format to the old format we were previously generating when all
tokens were JWTs and use the same JSON structure.

This code (and route) can be dropped once JupyterHub can switch to
the new API and there's a new UI to show users their login metadata.