Change scopes column to 512 characters
Assuming a scope length of about 13 characters, this should give
us 36 to 38 scopes, which hopefully will be more than enough.
Previously we'd get 18 or 19, which was probably already enough,
but better safe than sorry.

Ideally this would use a TEXT type, but unfortunately searches are
then slow and this has to be an indexed column to find appropriate
internal tokens.