Use react-alert for global alerts
Replace the internal ErrorBanner component with react-alert for
everyting except token creation and editing alerts.  Use the
default style, which is awful and converts everyting to all caps,
but we can fix that later.

Adjust some error parsing to account for the change in format of
422 error messages from the backend, and return a proper error
when redirecting for authentication.