Better error handling for token change search
Refactor token change search to move query parsing to the top-level
page and be cleaner about what data structures we pass around and
when we convert them.  Add structured error handling for backend
failures and use that to associate error messages with the correct
form field, using alerts for any errors we can't find a place for.