Parse error messages for token create and modify
Following the work for the token change search form, also parse
error messages in the token creation and editing modals and
associate errors with specific form fields, using alerts for
errors that cannot be associated.

This eliminates the last uses of the useError hook and the
ErrorBanner component, so remove them.