Add tox support for testing with PostgreSQL
Use the tox-docker add-on and support local testing with PostgreSQL.
This required some refactoring to change the database URL based on
the environment variables set by tox-docker.

Fix tests with IPv6 addresses to compare the canonicalized versions,
since PostgreSQL will canonicalize in the round-trip through the
database.

Use the same database session for SetupTest and its created factory,
remove the ability of a factory to create a session, and clean up
the session creation.  This allows closing the session cleanly when
ending the test, which is required to allow dropping the schema
between tests for PostgreSQL.

Set the collation to C on the token in the token table when using
PostgreSQL to retain the sorting that SQLite does by default.

Add a missing flush when creating subtokens so that they work
properly with a real database.

Unfortunately, tox-docker doesn't support pyproject.toml, so move
the tox settings to tox.ini.