Cascade expiration changes to subtokens
When the expiration of a user token is changed, this should affect
all of its subtokens if they would currently outlive the user token.
Detect this case and cascade changes to all subtokens.

As a side effect of this, also add a test that the Redis TTL on token
keys is set correctly.