Use nvm to manage Node in the Docker build
Rather than relying on the version of Node shipped with Debian,
install nvm as part of the dependencies build and use it to install
the pinned version of Node from .nvmrc.

This should fix build failures in CI because Gatsby wants a newer
version of Node than is available in the base Docker image.