Flesh out OpenAPI schema
Add lots of metadata to models, dependencies, and routes to make
the generated documentation more useful.  Refactor the
authentication dependencies so that the X-CSRF-Token header is not
part of the schema for read-only routes.  Add a model for error
messages, since FastAPI doesn't seem to export its.

Continue to try to keep handler code as thin as possible by moving
models that had accumulated there into gafaelfawr.models.  Use a
separate model for returning an InfluxDB token so that we can more
easily define a different example.

This updates all routes except the OpenID Connect routes.