Reorganize and document OpenID Connect support
Merge the userinfo and well-known handlers into the same file as
the rest of the OpenID Connect support since these are now only
used with OpenID Connect clients.  Move all of their models into
gafaelfawr.models.oidc.

Change the preferred URL of the userinfo endpoint to
/auth/openid/userinfo to keep all of the OpenID Connect support
together.  The old path will still work but is not advertised.

Change RSAKeyPair to return a JWKS model rather than a free-form
list of dicts, and adjust callers accordingly.  This simplifies the
handler code.

Add documentation and examples to all of the OpenID Connect routes
for auto-generated documentation purposes.