Rework how the UI is built
Use a three-stage workflow that first builds the UI and runs lints
and then stores the UI as a workflow artifact.  The testing then
uses that artifact to run the tests (without having to rebuild it),
and then the Docker image copies the same artifact into the Docker
image.  This removes all the Docker complexity of installing nvm
and doing Gatsby builds in the dependency container, at the cost of
making the Docker build a bit less self-contained.