Add support for maintaining service tokens
Add a configuration option to list service tokens to create in
Kubernetes, and add a new entry point and service to scan Kubernetes
for existing secrets with a gafaelfawr.lsst.io/token-type label,
create any secrets that should exist that are missing, update any
secrets that aren't valid, and delete any secrets that do not appear
in the configuration.

Move responsibility for creating fake token data for bootstrap and
internal tokens into the TokenData model and out of the
authentication dependency.

Allow the ip_address parameter on create_token_from_admin_request
to be None so that it can be used internally by Gafaelfawr background
jobs such as Kubernetes secret management.