Fix tests with real Redis and PostgreSQL
The new standalone factory poses some new challenges for thread
pools (particularly during testing) and for Redis connection pool
management.

Unwind the change to the Selenium tests to use the standalone
factory since it breaks a webapp in the same process.  Change the
CLI tests to avoid using SetupTest, since it is async but in a
separate thread pool from the one the command line code will use.

Avoid using fastapi_sqlalchemy in the standalone factory because it
requires an ASGI app, and close down the session and the Redis pool
correctly.

Rework the CLI tests a bit and create a separate setup function for
tests that cannot use SetupTest.