Allow separate database password configuration
Add a separate configuration setting for the database password and
fix support for configuration settings via environment variables
(which I thought was already working but which I didn't previously
test).  This allows the database password to be set via the
environment and thus in Kubernetes via a Secret, while still
generating the configuration file via a ConfigMap.