Re-add email support, expose full name
The Portal application uses the user's full name and email address.
Restore support to expose the full name in a response header from
the /auth route and add back in tracking and storage of the email
address.  Allow the email address to be set via the admin token
creation API.