Redo how Kubernetes secrets are handled
Rather than using special labels for managed secrets and adding
the list of secrets to create to the Gafaelfawr configuration,
add support for a GafaelfawrServiceToken custom resource.  Locate
all such resources and create corresponding secrets when the
update-service-tokens command is run.

This work uncovered that assert failures in the supporting test
libraries weren't being expanded into a comparison.  Fix that by
telling pytest that everything under tests is a test module.