Add --settings to update-service-tokens
Allow passing a settings file on the command line to
update-service-tokens to make it easier to test.

Fix the default for the init command to only override the config
path if the option was given.  This is necessary for a test to
work properly, and while we don't currently test init, make it
consistent with update-service-tokens so that we could.