Add Kubernetes controller invocation
Add support for watching a Kubernetes cluster for changes to our
custom resources and implementing appropriate changes to secrets
in reaction to seeing an event.

Unfortunately, the Kubernetes API only supports watching for one
type of object at a time, so use a background thread to do the
watching and a thread-safe queue to pass events to the main
process.  This pattern will allow watching multiple custom resource
types in the future, which will be desirable for renewable notebook
tokens and custom ingress processing.