Track generations in custom resource handling
Modifying the GafaelfawrServiceToken object to record the status
of the child secret triggers another MODIFY event.  To avoid
an infinite loop if the modification fails, remember the last
generation that we attempted to process and don't try to process
the object again for events indicating the same generation unless
there was an intervening delete.

Fix the mock and calls to replace_namespaced_secret to correctly
pass in the name.