Separate metadata updates from token updates
We want to propagate changes to annotations and labels from the
GafaelfaerServiceToken to the child Secret, but changes to those
attributes don't change the generation.  Therefore, handle them
separately and specially by comparing the metadata we care about
between the Secret and the GafaelfawrServiceToken on every event
other than a delete.