Fix version in built Docker image
setuptools_scm was generating a development version for the
Gafaelfawr installed in the Docker image instead of a proper
version based on version tags.  Fix this by not excluding
checked-in directories via .dockerignore (since they show up as
local modifications), not using /app as a working directory, and
checking out the tags when building Docker images.