Revoke GitHub OAuth grant on logout
During the first authentication from GitHub, the user chooses what
data to release to our application.  We have had persistent problems
where the user doesn't release the right data or they are not a
member of the correct organization when they first log on, and then
logging out and back in doesn't necessarily update what information
we can access via the GitHub access token.

On explicit logout, if the user logged on with GitHub, revoke the
OAuth authorization grant so that they're forced to go through the
permission screen again when they log in.  This should hopefully
avoid this issue at the cost of an extra screen when they return
after an explicit logout.