Return HTML errors on failure to log in
If the user failed to log in, show a human-readable error message
using an HTML template instead of returning a JSON message.  Use
a status code of 403 for all failures to log in except ones with
purely invalid URL syntax.

Additional HTML can be added to the error page by setting the
error_footer configuration option.

The HTML is currently unstyled.  Better styling will be added as
part of future UI work.