Send username as the full name
HTTP headers are not guaranteed to support character sets other
than ASCII, and Starlette forces them to ISO 8859-1.  This
interferes with correctly passing the user's full name to
protected services via HTTP headers.  For now, send the user's
username as their full name, since the username is guaranteed to
be ASCII.