Stop sending X-Auth-Request-Name
This was only used by the Portal, and the Portal falls back to
showing the username, which is what we want anyway.  We cannot
correctly represent human names in an HTTP header field without
further encoding, so stop trying to do so.  Protected applications
that need the name can request a delegated token and then get it
from the user-info API endpoint.