Refactor GitHub login tests
Extract the common pattern of performing a GitHub login into a
function and call it from the GitHub tests.

Move the test for alternate username and UID claims in JWTs to
the OIDC login test suite where it can be tested properly.  The
test with GitHub became a no-op once we switched to non-JWT
tokens.